refactored

+++ engine/components/position2d.go
@@ -1,8 +1,8 @@
 package components
 
 type Position2D struct {
-	xPos float64
-	yPos float64
+	xPos float32
+	yPos float32
 }
 
 func NewPosition2DComponent() *Position2D {
@@ -12,22 +12,22 @@ func NewPosition2DComponent() *Position2D {
 	}
 }
 
-func BuildPosition2DComponent(xPos float64, yPos float64) *Position2D {
+func BuildPosition2DComponent(xPos float32, yPos float32) *Position2D {
 	return &Position2D{
 		xPos: xPos,
 		yPos: yPos,
 	}
 }
 
-func (node *Position2D) XPos() float64 {
+func (node *Position2D) XPos() float32 {
 	return node.xPos
 }
 
-func (node *Position2D) YPos() float64 {
+func (node *Position2D) YPos() float32 {
 	return node.yPos
 }
 
-func (node *Position2D) SetPosition2D(xPos float64, yPos float64) {
+func (node *Position2D) SetPosition2D(xPos float32, yPos float32) {
 	node.xPos = xPos
 	node.yPos = yPos
 }
